refac: rmv category and change in video card
