--- src/Components/Categories/Categories.module.css
@@ -1,61 +1,0 @@
-.section_heading {
-  text-align: center;
-  padding: 2rem 0 3rem 0;
-  font-size: 2.8rem;
-  color: var(--category-color);
-  text-decoration: underline;
-  text-decoration-color: var(--retro-color);
-}
-
-.category_section {
-  padding: 3rem 0;
-}
-
-.link {
-  text-decoration: none;
-}
-
-.category_container {
-  display: flex;
-  justify-content: space-around;
-  row-gap: 1rem;
-}
-
-.category_col {
-  position: relative;
-  max-width: 300px;
-  height: 28rem;
-}
-
-.category_col .after {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: #fff;
-  font-size: 1.8rem;
-  font-weight: bold;
-  letter-spacing: 1px;
-  visibility: hidden;
-  transition: 0.2s ease-in-out;
-}
-
-.category_pic {
-  max-width: 100%;
-  height: 28rem;
-}
-
-.category_col:hover .after {
-  visibility: visible;
-  background: rgba(255, 6, 6, 0.5);
-}
-
-@media only screen and (max-width: 1024px) {
-  .category_container {
-    flex-wrap: wrap;
-  }
-}
