+++ src/Pages/Home/Home.module.css
@@ -1,8 +1,3 @@
-.hero_container {
-  position: relative;
-  width: 100%;
-}
-
 .logo {
   font-family: "Carter One", cursive;
   font-size: 1.4rem;
@@ -40,11 +35,15 @@
   border-radius: 50%;
 }
 
-.hero_video:before {
-  content: "";
-  position: absolute;
-  height: 100%;
-  width: 100%;
+.hero_description_container {
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+  position: relative;
+  min-height: 100vh;
+  overflow: hidden;
   background-image: linear-gradient(
     to top,
     rgba(0, 0, 0, 0),
@@ -52,15 +51,19 @@
   );
 }
 
-.hero_description_container {
-  max-width: 620px;
+.hero_text {
   position: absolute;
-  z-index: 5;
-  margin: 0 auto;
-  top: 40%;
-  left: 0;
-  right: 0;
   text-align: center;
+  z-index: 5;
+}
+
+.hero_video {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  min-width: 100%;
+  min-height: 100%;
 }
 
 .hero_title {
